feat(product): add register product logic

+++ src/main/kotlin/com/lambsroad/trendella/modules/product/application/port/CategoryDeletePort.kt
@@ -1,3 +1,3 @@
 package com.lambsroad.trendella.modules.product.application.port
 
-data class CategoryDeletePort(val categoryId: Int)
+data class CategoryDeletePort(val categoryId: Long)
